pyproject.toml
[project]
name = "simple-analytics"
version = "0.1.0"
description = "Flexible multi-project analytics platform with Django and React"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.7.0",
    "django-extensions>=4.1",
    "django-rq>=3.1",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
]

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "black>=25.1.0",
    "coverage>=7.6.0",
    "django-stubs>=5.2.2",
    "factory-boy>=3.3.0",
    "faker>=30.0.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "mypy-extensions>=1.1.0",
    "pre-commit>=4.3.0",
    "pytest>=8.3.0",
    "pytest-django>=4.9.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.12.12",
    "safety>=3.6.1",
]

# Tool configurations
[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "DJ",  # flake8-django
]
ignore = [
    "E501",  # line too long (handled by black)
    "DJ001", # Avoid using `null=True` on string-based fields - we use it intentionally
]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
    "venv",
    "env",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"settings.py" = ["F405", "F403"]  # Allow star imports in settings
"*/migrations/*.py" = ["E501"]  # Allow long lines in migrations

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = "django"
known_first_party = ["projects", "events", "analytics_backend"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = ["migrations", ".venv", "venv"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = false  # Django models don't need explicit typing
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = false
exclude = [
    "migrations",
    "venv",
    ".venv",
]

# Django-specific settings
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "analytics_backend.settings"

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.bandit]
exclude_dirs = ["tests", "migrations", ".venv"]
skips = ["B101", "B311"]  # Skip assert_used and random generator tests (sampling is not cryptographic)

[tool.bandit.assert_used]
skips = ["*test*.py", "*/test_*.py"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "analytics_backend.settings"
python_files = ["tests.py", "test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--disable-warnings",
    "--reuse-db",
    "--tb=short"
]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with -m 'not slow')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["projects", "events"]
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/.venv/*",
    "*/tests/*",
    "manage.py",
    "analytics_backend/wsgi.py",
    "analytics_backend/asgi.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]
